unify platform and collection layout
