+++ frontend/src/components/Gallery/AppBar/Collection/Base.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import FilterBtn from "@/components/Gallery/AppBar/common/FilterBtn.vue";
+import CollectionInfoDrawer from "@/components/Gallery/AppBar/Collection/CollectionInfoDrawer.vue";
 import FilterDrawer from "@/components/Gallery/AppBar/common/FilterDrawer/Base.vue";
 import FilterTextField from "@/components/Gallery/AppBar/common/FilterTextField.vue";
 import GalleryViewBtn from "@/components/Gallery/AppBar/common/GalleryViewBtn.vue";
@@ -11,6 +12,7 @@ import storeRoms from "@/stores/roms";
 import type { Events } from "@/types/emitter";
 import type { Emitter } from "mitt";
 import { storeToRefs } from "pinia";
+import storeNavigation from "@/stores/navigation";
 import { inject, ref } from "vue";
 import { useDisplay } from "vuetify";
 
@@ -21,29 +23,9 @@ const viewportWidth = ref(window.innerWidth);
 const auth = storeAuth();
 const romsStore = storeRoms();
 const { currentCollection } = storeToRefs(romsStore);
+const navigationStore = storeNavigation();
+const { activeCollectionInfoDrawer } = storeToRefs(navigationStore);
 const open = ref(false);
-const collectionInfoFields = [
-  {
-    key: "name",
-    label: "Name",
-  },
-  {
-    key: "description",
-    label: "Description",
-  },
-  {
-    key: "rom_count",
-    label: "Roms",
-  },
-  {
-    key: "user__username",
-    label: "Owner",
-  },
-  {
-    key: "is_public",
-    label: "Public",
-  },
-];
 </script>
 
 <template>
@@ -58,9 +40,9 @@ const collectionInfoFields = [
   >
     <template #prepend>
       <v-btn
-        :color="open ? 'romm-accent-1' : ''"
+        :color="activeCollectionInfoDrawer ? 'romm-accent-1' : ''"
         rounded="0"
-        @click="open = !open"
+        @click="navigationStore.swtichActiveCollectionInfoDrawer"
         icon="mdi-information"
       ></v-btn>
       <filter-btn />
@@ -72,91 +54,7 @@ const collectionInfoFields = [
     </template>
   </v-app-bar>
 
-  <v-navigation-drawer
-    v-model="open"
-    floating
-    mobile
-    :width="xs ? viewportWidth : '500'"
-    v-if="currentCollection"
-  >
-    <v-row no-gutters class="text-center justify-center align-center mt-4">
-      <v-col style="max-width: 240px">
-        <collection-card
-          :key="currentCollection.updated_at"
-          :collection="currentCollection"
-        />
-        <v-btn
-          rounded="4"
-          @click="
-            emitter?.emit('showEditCollectionDialog', { ...currentCollection })
-          "
-          class="mt-4 bg-terciary"
-        >
-          <template #prepend>
-            <v-icon>mdi-pencil-box</v-icon>
-          </template>
-          Edit collection
-        </v-btn>
-      </v-col>
-    </v-row>
-    <v-row no-gutters class="mt-4">
-      <v-col cols="12">
-        <v-card class="mx-4 bg-terciary" elevation="0">
-          <v-card-text class="pa-4">
-            <template
-              v-for="(field, index) in collectionInfoFields"
-              :key="field.key"
-            >
-              <div
-                v-if="
-                  currentCollection[field.key as keyof typeof currentCollection]
-                "
-                :class="{ 'mt-4': index !== 0 }"
-              >
-                <p class="text-subtitle-1 text-decoration-underline">
-                  {{ field.label }}
-                </p>
-                <p class="text-subtitle-2">
-                  {{
-                    currentCollection[
-                      field.key as keyof typeof currentCollection
-                    ]
-                  }}
-                </p>
-              </div>
-            </template>
-          </v-card-text>
-        </v-card>
-      </v-col>
-    </v-row>
-    <v-row no-gutters class="mt-4">
-      <v-col cols="12">
-        <r-section
-          v-if="auth.scopes.includes('collections.write')"
-          icon="mdi-alert"
-          icon-color="red"
-          title="Danger zone"
-          elevation="0"
-        >
-          <template #content>
-            <div class="text-center my-2">
-              <v-btn
-                class="text-romm-red bg-terciary"
-                variant="flat"
-                @click="
-                  emitter?.emit('showDeleteCollectionDialog', currentCollection)
-                "
-              >
-                <v-icon class="text-romm-red mr-2">mdi-delete</v-icon>
-                Delete collection
-              </v-btn>
-            </div>
-          </template>
-        </r-section>
-      </v-col>
-    </v-row>
-  </v-navigation-drawer>
-
+  <collection-info-drawer />
   <filter-drawer />
 </template>
