+++ frontend/src/components/Gallery/AppBar/Collection/CollectionInfoDrawer.vue
@@ -1,0 +1,139 @@
+<script setup lang="ts">
+import FilterBtn from "@/components/Gallery/AppBar/common/FilterBtn.vue";
+import FilterDrawer from "@/components/Gallery/AppBar/common/FilterDrawer/Base.vue";
+import FilterTextField from "@/components/Gallery/AppBar/common/FilterTextField.vue";
+import GalleryViewBtn from "@/components/Gallery/AppBar/common/GalleryViewBtn.vue";
+import SelectingBtn from "@/components/Gallery/AppBar/common/SelectingBtn.vue";
+import CollectionCard from "@/components/common/Collection/Card.vue";
+import RSection from "@/components/common/RSection.vue";
+import storeAuth from "@/stores/auth";
+import storeRoms from "@/stores/roms";
+import storeNavigation from "@/stores/navigation";
+import type { Events } from "@/types/emitter";
+import type { Emitter } from "mitt";
+import { storeToRefs } from "pinia";
+import { inject, ref } from "vue";
+import { useDisplay } from "vuetify";
+
+// Props
+const { xs } = useDisplay();
+const emitter = inject<Emitter<Events>>("emitter");
+const viewportWidth = ref(window.innerWidth);
+const auth = storeAuth();
+const romsStore = storeRoms();
+const { currentCollection } = storeToRefs(romsStore);
+const navigationStore = storeNavigation();
+const { activeCollectionInfoDrawer } = storeToRefs(navigationStore);
+const collectionInfoFields = [
+  {
+    key: "name",
+    label: "Name",
+  },
+  {
+    key: "description",
+    label: "Description",
+  },
+  {
+    key: "rom_count",
+    label: "Roms",
+  },
+  {
+    key: "user__username",
+    label: "Owner",
+  },
+  {
+    key: "is_public",
+    label: "Public",
+  },
+];
+</script>
+
+<template>
+  <v-navigation-drawer
+    v-model="activeCollectionInfoDrawer"
+    floating
+    mobile
+    :width="xs ? viewportWidth : '500'"
+    v-if="currentCollection"
+  >
+    <v-row no-gutters class="justify-center align-center pa-4">
+      <v-col style="max-width: 240px" cols="12">
+        <div class="text-center justify-center align-center">
+          <collection-card
+            :key="currentCollection.updated_at"
+            :collection="currentCollection"
+          />
+        </div>
+        <div
+          class="text-center mt-4"
+          v-if="auth.scopes.includes('collections.write')"
+        >
+          <v-btn
+            rounded="4"
+            @click="
+              emitter?.emit('showEditCollectionDialog', {
+                ...currentCollection,
+              })
+            "
+            class="bg-terciary"
+          >
+            <template #prepend>
+              <v-icon>mdi-pencil-box</v-icon>
+            </template>
+            Edit collection
+          </v-btn>
+        </div>
+      </v-col>
+      <v-col cols="12">
+        <v-card class="mt-4 bg-terciary fill-width" elevation="0">
+          <v-card-text class="pa-4">
+            <template
+              v-for="(field, index) in collectionInfoFields"
+              :key="field.key"
+            >
+              <div
+                v-if="
+                  currentCollection[field.key as keyof typeof currentCollection]
+                "
+                :class="{ 'mt-4': index !== 0 }"
+              >
+                <p class="text-subtitle-1 text-decoration-underline">
+                  {{ field.label }}
+                </p>
+                <p class="text-subtitle-2">
+                  {{
+                    currentCollection[
+                      field.key as keyof typeof currentCollection
+                    ]
+                  }}
+                </p>
+              </div>
+            </template>
+          </v-card-text>
+        </v-card>
+      </v-col>
+    </v-row>
+    <r-section
+      v-if="auth.scopes.includes('collections.write')"
+      icon="mdi-alert"
+      icon-color="red"
+      title="Danger zone"
+      elevation="0"
+    >
+      <template #content>
+        <div class="text-center">
+          <v-btn
+            class="text-romm-red bg-terciary ma-2"
+            variant="flat"
+            @click="
+              emitter?.emit('showDeleteCollectionDialog', currentCollection)
+            "
+          >
+            <v-icon class="text-romm-red mr-2">mdi-delete</v-icon>
+            Delete collection
+          </v-btn>
+        </div>
+      </template>
+    </r-section>
+  </v-navigation-drawer>
+</template>
